add user file permission hook to classroom code interface
