+++ src/components/classrooms/ClassroomCodeInterface.tsx
@@ -1,6 +1,8 @@
+import useUpdateUserFilePermissions from '../../hooks/useUpdateUserFilePermissions';
 import { CodeInterface } from '../CodeInterface/CodeInterface';
 
 export default function ClassroomCodeInterface() {
+  useUpdateUserFilePermissions();
   return (
     <div className="h-full">
       <CodeInterface className="h-full" />
